+++ src/L2dotNET/Network/serverpackets/ExBuySellListSell.cs
@@ -34,7 +34,7 @@ namespace L2dotNET.Network.serverpackets
                 WriteShort(item.Template.Type2);
                 WriteShort(item.CustomType1);
                 WriteShort(0);
-                WriteInt(item.Template.BodyPart);
+                WriteInt((int) item.Template.BodyPart);
                 WriteShort(item.Enchant);
                 WriteShort(item.CustomType2);
                 WriteInt(item.AugmentationId);
@@ -59,7 +59,7 @@ namespace L2dotNET.Network.serverpackets
                 WriteShort(item.Template.Type2);
                 WriteShort(item.CustomType1);
                 WriteShort(0);
-                WriteInt(item.Template.BodyPart);
+                WriteInt((int) item.Template.BodyPart);
                 WriteShort(item.Enchant);
                 WriteShort(item.CustomType2);
                 WriteInt(item.AugmentationId);
